Fix userController

+++ src/main/java/antonioloiacono/tesi/monolitica/service/imp/UserServiceImpl.java
@@ -1,23 +1,55 @@
 package antonioloiacono.tesi.monolitica.service.imp;
 
+import antonioloiacono.tesi.monolitica.dto.UserSaveDTO;
+import antonioloiacono.tesi.monolitica.dto.UserUpdateDTO;
 import antonioloiacono.tesi.monolitica.entity.User;
+import antonioloiacono.tesi.monolitica.exception.ResourceAlreadyExistsException;
+import antonioloiacono.tesi.monolitica.exception.ResourceNotFoundException;
 import antonioloiacono.tesi.monolitica.repository.UserRepository;
 import antonioloiacono.tesi.monolitica.service.UserService;
+import antonioloiacono.tesi.monolitica.util.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
+    private UserRepository userRepository;
+    private ModelMapper modelMapper;
 
     @Autowired
-    UserRepository userRepository;
+    public UserServiceImpl(UserRepository userRepository, ModelMapper modelMapper) {
+        this.userRepository = userRepository;
+        this.modelMapper = modelMapper;
+    }
 
     @Override
-    public User saveUser(User user) {
+    public User saveUser(UserSaveDTO dto) {
+        String email = dto.getEmail();
+        if (userRepository.existsByEmail(email)) {
+            throw new ResourceAlreadyExistsException("User already exist with email: " + email);
+        }
+        return userRepository.save(modelMapper.map(dto, User.class));
+    }
+
+    @Override
+    public User updateUser(Long id, UserUpdateDTO dto) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()) {
+            throw new ResourceNotFoundException("No user to update found with the id: " + id);
+        }
+        User user = optionalUser.get();
+        if (dto.getEmail() != null){
+            user.setEmail(dto.getEmail());
+        }
+        if (dto.getFirstName() != null){
+            user.setFirstName(dto.getFirstName());
+        }
+        if (dto.getLastName() != null){
+            user.setLastName(dto.getLastName());
+        }
         return userRepository.save(user);
     }
 
@@ -29,12 +61,20 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<User> findUserById(Integer id) {
-        return userRepository.findById(id);
+    public User findUserById(Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()) {
+            throw new ResourceNotFoundException("No user found with the id: " + id);
+        }
+        return optionalUser.get();
     }
 
     @Override
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()) {
+            throw new ResourceNotFoundException("No user to delete found with the id: " + id);
+        }
         userRepository.deleteById(id);
     }
